+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
 - Setting to ignore traps and filler items when connecting, enabled by default.
 - In key item gating mode, the closed doors will have a lock icons instead of the "no entry" symbol.
 - Updated setup guide to include linux/steam deck information.
+- When entering a level with ramdom level entrnace, the alternative entrances to that level will be revealed on the map.
+  - This applies to Speed Highway, Lost World, Final Egg and Sky Deck.
 
 ### Changed: